@@ -14,19 +14,19 @@ public class TileLayerFormatOptionTests
     {
         // Arrange
         var command = new Command("test");
-        var option = new TileLayerFormatOption();
+        var tileLayerFormatOption = new TileLayerFormatOption();
 
         // Act
-        option.Register(command);
+        tileLayerFormatOption.Register(command);
         var result = command.Parse($"--format {format}");
 
         // Assert
         Assert.Empty(result.Errors);
+        Assert.Equal(format, result.GetValueForOption(tileLayerFormatOption.Option));
     }
 
     [Theory]
     [InlineData("")]
-    [InlineData(null)]
     [InlineData("unknown")]
     [InlineData("bzip2")]
     public void Register_InvalidFormats_SetErrorMessage(string format)
