@@ -1,9 +1,8 @@
 ﻿using System.CommandLine;
-using Animation2Tilemap.CommandLineOptions.Contracts;
 
 namespace Animation2Tilemap.CommandLineOptions;
 
-public class OutputOption : ICommandLineOption<string>
+public class OutputOption
 {
     public OutputOption()
     {
