@@ -1,4 +1,7 @@
-﻿namespace Animation2Tilemap.Services.Contracts;
+﻿using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
+
+namespace Animation2Tilemap.Services.Contracts;
 
 public interface IImageHashService
 {
